[1.6.x] Add a patch level to the Django version.
These security backports will be released with a 1.6.11.p identifier,
where .p is a patch level. We'll start with 1 (giving us 1.6.11.1) and
go from there.

This is used instead of local version identifiers (like "+beanbag.1")
for compatibility and dependency expression reasons.

Along with this, I've dropped the version down from 1.6.12 alpha 0 to
1.6.11, since we're dealing explicitly with security releases to that
version.